feat: 🎸 reduce inquirer
