+++ packages/create-pkg/package.json
@@ -21,11 +21,9 @@
     "typescript": "^4.5.4"
   },
   "dependencies": {
-    "@appworks/cli": "^3.0.1",
     "@iceworks/generate-material": "^1.0.12",
     "cac": "^6.7.12",
     "consola": "^2.15.3",
-    "decamelize": "^6.0.0",
     "fs-extra": "^10.0.0",
     "ice-npm-utils": "^3.0.2",
     "inquirer": "^8.2.1",
