--- packages/create-pkg/src/generateNpmName.ts
@@ -1,7 +1,0 @@
-import decamelize from 'decamelize';
-
-export function generateNpmName(name: string, npmScope?: string): string {
-  // WebkitTransform -> webkit-transform
-  const decamelizeName = decamelize(name, { separator: '-' });
-  return npmScope ? `${npmScope}/${decamelizeName}` : decamelizeName;
-}
